+++ src/androidTest/kotlin/us/frollo/frollosdk/base/ResourceTest.kt
@@ -32,7 +32,7 @@ class ResourceTest {
 
     @Before
     fun setup() {
-        FrolloSDK.app = app
+        FrolloSDK.context = app
     }
 
     @Test