+++ src/androidTest/kotlin/us/frollo/frollosdk/error/DataErrorTest.kt
@@ -29,7 +29,7 @@ class DataErrorTest {
 
     @Before
     fun setUp() {
-        FrolloSDK.app = app
+        FrolloSDK.context = app
     }
 
     @Test