+++ src/androidTest/kotlin/us/frollo/frollosdk/error/FrolloSDKErrorTest.kt
@@ -30,7 +30,7 @@ class FrolloSDKErrorTest {
 
     @Before
     fun setUp() {
-        FrolloSDK.app = app
+        FrolloSDK.context = app
     }
 
     @Test