+++ src/androidTest/kotlin/us/frollo/frollosdk/error/NetworkErrorTest.kt
@@ -30,7 +30,7 @@ class NetworkErrorTest {
 
     @Before
     fun setUp() {
-        FrolloSDK.app = app
+        FrolloSDK.context = app
     }
 
     @Test