+++ src/androidTest/kotlin/us/frollo/frollosdk/error/OAuth2ErrorTest.kt
@@ -32,7 +32,7 @@ class OAuth2ErrorTest {
 
     @Before
     fun setUp() {
-        FrolloSDK.app = app
+        FrolloSDK.context = app
     }
 
     @Test