+++ src/androidTest/kotlin/us/frollo/frollosdk/model/coredata/aggregation/providers/ProviderLoginFormTest.kt
@@ -45,7 +45,7 @@ class ProviderLoginFormTest {
 
     @Before
     fun setUp() {
-        FrolloSDK.app = app
+        FrolloSDK.context = app
     }
 
     @Test