+++ changelog.md
@@ -6,6 +6,7 @@
 
 - Optimize Transactions advanced filtering and searching
 - Add Account Features & CDR Product Information to Account API Response
+- Migrate from Application Context to ContentProvider Context
 
 ### 3.5.0